[CORE] Backport crypto libraries.

+++ src/crypto/ripemd160.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2014 The Bitcoin developers
+// Copyright (c) 2014-2016 The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -12,17 +12,17 @@
 class CRIPEMD160
 {
 private:
-    uint32_t s[5];
-    unsigned char buf[64];
-    size_t bytes;
+  uint32_t s[5];
+  unsigned char buf[64];
+  uint64_t bytes;
 
 public:
-    static const size_t OUTPUT_SIZE = 20;
+  static const size_t OUTPUT_SIZE = 20;
 
-    CRIPEMD160();
-    CRIPEMD160& Write(const unsigned char* data, size_t len);
-    void Finalize(unsigned char hash[OUTPUT_SIZE]);
-    CRIPEMD160& Reset();
+  CRIPEMD160();
+  CRIPEMD160& Write(const unsigned char* data, size_t len);
+  void Finalize(unsigned char hash[OUTPUT_SIZE]);
+  CRIPEMD160& Reset();
 };
 
 #endif // BITCOIN_CRYPTO_RIPEMD160_H
